devops/requirements.txt
django==3.1.5
drf-yasg==1.20.0
django-cors-headers==3.6.0
django-filter==2.4.0
djangorestframework==3.12.2
PyJWT==1.7.1
djangorestframework-simplejwt==4.6.0
Pillow>=8.0.1
python-decouple==3.3
psycopg2-binary==2.8.6
unidecode==1.1.2
coreapi==2.3.3
pyyaml==5.3.1
requests>=2.25.1
coverage==5.3.1
image-helpers==0.1.4
validate-docbr==1.8.2
drf-flex-fields==0.8.9
django-ordered-model==3.4.3
python-dateutil==2.8.1
workalendar==15.4.0
drf-api-tracking==1.8.0
dictdiffer==0.9.0
python-telegram-bot==13.7.0

